Allow parsing `pre` on contained items
This means that for nested `pre` attributes only the outmost one is
actually processed, the rest will be handled during that processing.

This will allow a single `#![pre::pre]` to be put at the top of the
file, once that is stabilized. (It currently doesn't work, because rustc
passes an invalid module without a name to the macro. However it can be
used on an outer module, to apply it to the whole module.)

This is not very useful as of yet, but it will be useful, once
`#[pre::check_pre]` is replaced by `#[pre::pre]`.