Remove unnecessary `#[allow(dead_code)]`
This does not seem to trigger in this case.
Even if it will trigger in the future, it would probably trigger an
`unreachable_code` lint instead.
Also it is not clear whether the whole if expression is the best place
to put the attribute. Therefore it will be removed for now and can be
re-added, if the problem surfaces.