Always format `assure` on multiple lines
This expanded form should be more readable in most situations.
Changing it at every place in the documentation hopefully encourages the
use of this form in the code of users.

I decided not to add a dedicated paragraph encouraging this use, as I
think this is a question of style and `pre` should not be too
opinionated about style.

Tests were not changed, as those mostly aren't very representative of
real code anyway and are rather short.